requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tweepy>=4.14.0
praw>=7.7.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
pyyaml>=6.0
openai>=1.0.0
